requirements.txt
# --- Core Python utils ---
numpy>=1.24,<2.0
tqdm>=4.66

# --- Document loading ---
pypdf>=4.0

# --- Embeddings ---
sentence-transformers==2.7.0
transformers==4.41.2
torch>=2.1.0

# --- Vector store ---
faiss-cpu>=1.7.4

# --- HTTP client (Ollama API) ---
requests>=2.31

# --- API server ---
fastapi>=0.110
uvicorn>=0.29


python-multipart